Change word to use int as base

+++ include/c64/types.h
@@ -2,7 +2,7 @@
 #define C64_TYPES_H
 
 typedef unsigned char	byte;
-typedef unsigned short	word;
+typedef unsigned int	word;
 typedef unsigned long	dword;
 
 typedef signed char	sbyte;
